fix: rename mock file to ensure compatibility
