+++ src/__tests__/eventListenerCallbacks.spec.ts
@@ -14,7 +14,7 @@ import {DEVICE_CONNECTIVITY_EVENT} from '../internal/privateTypes';
 // Mock modules
 require('jest-fetch-mock').enableMocks();
 jest.mock('./../internal/nativeModule', () =>
-  require('./../../jest/nativeModule.mock'),
+  require('./../../jest/netinfo-mock'),
 );
 
 describe('react-native-community/netinfo', () => {
